requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram_calendar==0.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.0
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.3
multidict==6.4.4
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.13.2
yarl==1.20.0

XlsxWriter~=3.2.3


pdfplumber==0.11.5
pdfminer.six==20231228